Consumable items now can be used by clicking on them in inventory with LMB
* Equipment with lvl > player's lvl now can't be used (except for lvl 1 items when player is at lvl 0)
* Map.addStructure(name, x, y) -- add map structures to specific coordinates
* Reduced Exp requirement for Player's level
* Non-human NPCs now can't give quests to player
* Consumables no longer replace currently held item on use

+++ MadSand-core/src/hitonoriol/madsand/entities/Entity.java
@@ -49,9 +49,9 @@ public abstract class Entity {
 	@JsonIgnore
 	public int movespeed = 2; // on-screen move speed (for smooth movement)
 	@JsonIgnore
-	public int stepy = MadSand.TILESIZE;
+	public float stepy = MadSand.TILESIZE;
 	@JsonIgnore
-	public int stepx = MadSand.TILESIZE;
+	public float stepx = MadSand.TILESIZE;
 
 	boolean stepping = false;
